requirements.txt
Flask==3.0.1
numpy==1.26.3
pandas==2.2.0
Pillow==10.2.0
uvicorn[standard]==0.22.0